students/K33402/Komarov_Georgy/LAB2/elevennote/requirements/base.txt
Django~=3.0.6
psycopg2-binary
python-decouple
pytz
uWSGI
django-widget-tweaks
django-wysiwyg
djangorestframework
djangorestframework-jwt
